Make the memory limit a megabyte count instead of a string

+++ Apps/AppManager/Source/Malterlib_Cloud_App_AppManager_EnvironmentVM.cpp
@@ -187,7 +187,7 @@ namespace NMib::NCloud::NAppManager
 		CreateParams.m_BundleDirectory = BundleDirectory;
 		CreateParams.m_RestoreImagePath = RestoreImage;
 		CreateParams.m_CPUCount = (uint32)_Params["VMCPUCount"].f_AsInteger();
-		CreateParams.m_MemoryMB = (uint64)_Params["VMMemoryMB"].f_AsInteger();
+		CreateParams.m_MemoryMB = (uint32)_Params["VMMemoryMB"].f_AsInteger();
 		CreateParams.m_DiskSizeGB = (uint64)_Params["DiskSizeGB"].f_AsInteger();
 
 		co_await _pCommandLine->f_StdOut("Creating VM image '{}' and installing macOS from {}\n"_f << Name << RestoreImage);
